% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.general-package.R
\docType{package}
\name{package.general-package}
\alias{package.general}
\alias{package.general-package}
\title{package.general: Fonctions d'intérêt collectif pour FA}
\description{
Différentes fonctions utiles de manipulation de données, de création de cartes, etc...
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
